voodoo: Add idle audio playback, movement threshold & cleanup
If audio has been idle for 10 seconds then we play a canned
audio file (with talking skull).

I also refactored some code in here to clean things up.

I also added a threshold for the minimum volume level to make
the mouth move at all so it doesn't start chattering if there
is a tiny sound.