fart-blaster: make light actions into state machines
This lets me run the main chamber and the bubble independently
with a single thread.  The advantage of doing so is that I don't
have to either have 2 NeoPixelPico objects or synchronize between
2 threads sharing a single object.

Note: I had to restrict the total number of lights lit by the
purlsing because the power draw is higher than the pi5 USB port
can provide.