mushrooms: tmc2208 hacks
I think that the UART was used to program different settings for
the two boards I'm using.

It seems like one board is running at 1/64 microstepping so I have
a jumper to cut the speed in half.  I also have some commented out
code for reversing the motor direction via a jumper because the
first board was inverted from the drv8825.